+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Bookinsanity font as body for export to PDF
 - `.two-column` div style
+- `.flavortext` blockquote style
 
 ### Fixed
 - Fixed same-order page sort being reversed when exporting to PDF